requirements.txt
attrs==22.2.0
autopep8==2.0.1
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
datalite==0.7.2
flake8==6.0.0
idna==3.4
iniconfig==2.0.0
mccabe==0.7.0
packaging==23.0
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.2.1
python-dotenv==1.0.0
requests==2.28.2
soupsieve==2.4
types-beautifulsoup4==4.11.6.6
types-html5lib==1.1.11.11
types-requests==2.28.11.13
types-urllib3==1.26.25.6
urllib3==1.26.14
